requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf == 6.33.1
scikit-learn
pandas
joblib
